+++ core/domain/order/order_manager.go
@@ -215,7 +215,7 @@ func (t *orderManagerImpl) applyCoupon(m member.IMemberAggregateRoot, o order.IO
 	po := py.Get()
 	//todo: ?? 重构
 	cp := t.promRepo.GetCouponByCode(
-		m.GetAggregateRootId(), couponCode)
+		int64(m.GetAggregateRootId()), couponCode)
 	// 如果优惠券不存在
 	if cp == nil {
 		return errors.New("优惠券无效")
@@ -225,13 +225,13 @@ func (t *orderManagerImpl) applyCoupon(m member.IMemberAggregateRoot, o order.IO
 	result, err := coupon.CanUse(m, float32(po.TotalAmount/100))
 	if result {
 		if coupon.CanTake() {
-			_, err = coupon.GetTake(m.GetAggregateRootId())
+			_, err = coupon.GetTake(int64(m.GetAggregateRootId()))
 			//如果未占用，则占用
 			if err != nil {
-				err = coupon.Take(m.GetAggregateRootId())
+				err = coupon.Take(int64(m.GetAggregateRootId()))
 			}
 		} else {
-			_, err = coupon.GetBind(m.GetAggregateRootId())
+			_, err = coupon.GetBind(int64(m.GetAggregateRootId()))
 		}
 		if err != nil {
 			domain.HandleError(err, "domain")
